+++ CHANGELOG.md
@@ -8,6 +8,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.1.2, 12/20/2024
+
+### Added
+
+Unit test in event-script-engine to validate that the config management system can merge
+parameters in application.yml and application.properties.
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 4.1.1, 12/18/2024
 
 ### Added